requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8
python-dotenv==1.0.0
schedule==1.2.1
lxml==5.1.0
lxml_html_clean==0.1.0
requests==2.31.0
